+++ apps/rig_outbound_gateway/lib/rig_outbound_gateway/kafka.ex
@@ -1,0 +1,10 @@
+defmodule RigOutboundGateway.Kafka do
+  @moduledoc """
+  Interface to Kafka-related functionality.
+  """
+
+  @type topic :: String.t()
+  @type partition :: non_neg_integer()
+
+  defdelegate ready?(topic, partition, timeout \\ 0), to: RigOutboundGateway.Kafka.Readiness
+end